Pointed out by Joerg Mayer:
The LDP dissector uses a pointer-to-func for the translation of addresses
to text. Replace the void * and obscure type casts with a storage class
definition for such a function, and type cast its argument in the cases
where it is not a const guint8 * anymore (e.g., IPv6 address).

Add a default type-casting function for unhandled cases.

Fix signedness complaint of the TCP and UDP port preferences for LDP.

svn path=/trunk/; revision=9634