"tmp_tag_text" can hold a string of up to 255 characters, which means it needs to be 256 characters long to hold a maximum-length string plus a terminating '\0', as noted by Stefan Esser.
Don't bother putting in the null terminator at the end in the case where
we're putting an IP address or address/port into "tmp_tag_text" -
"sprintf()" does that for you.

In the case where it's an IP address and port, do it all in one sprintf
call - I don't think there's a guarantee that

	sprintf(tmp_tag_text, "%s:%u", tmp_tag_text, ...)

works, although it could work.

Also, handle the case where the length is neither 4 (IP address) nor 6
(IP address/port).

svn path=/trunk/; revision=10418