Note that GTK+ 2.4 doesn't work, and indicate why.
Enumerate the dependencies for GLib 2.4/GTK+ 2.4, and indicate that GNU
gettext is also required, even though it's not in the set of
dependencies on the GTK+ site (after all, the ENTIRE WORLD runs Linux,
RIGHT?).

Give a (partial) order for the dependencies.

Note some issues with building and installing the JPEG and libpng
libraries.

svn path=/trunk/; revision=10443