Add a "report_failure()" routine to allow dissectors to report arbitrary errors to the user.  Use that, rather than "g_warning()", in the Diameter dissector to report errors reading the dictionary.
Make the format argument to "simple_dialog()" a "const" pointer.

Fix up the read-error message in Tethereal to end with a newline.

If a simple dialog is requested before the main window or the
capture-control window is popped up, queue it up and pop the queued
messages up once the main or capture-control window is displayed.

svn path=/trunk/; revision=10616