Make "little_endian" local to the routines to parse submessages.
Before checking for the "RTPS" tag, make sure we have at least 4 bytes
in the message.

Use "tvb_reported_length()" to get the length of the message, so we
throw an exception on a packet cut short by a snapshot length.

Put in a comment asking whether the byte order of the "offset to next
submessage" field is little-endian or specified by the E bit.

svn path=/trunk/; revision=10636