You can hand a string to "col_add_str()" as an argument - or to "col_set_str()", for that matter.
Eliminate a compiler warning by initializing a variable.

Don't use "match_strval()" on values you don't know for certain are in
the value_string table - use "val_to_str()", so that you don't get a
null pointer if the value isn't in that table.  ("Know for certain"
meaning "the code has made sure it's a value that's in the table", not
"the protocol spec says the field must have one of the values in the
table".)

svn path=/trunk/; revision=10859