Make the definition of "add_hostlist_table_data()" match its declaration.
Give the address and port arguments to "add_hostlist_table_data()"
appropriate names and Doxygen comments (they're not a source address and
port, they're just an address and port), and do the same for the address
and port members of a "hostlist_talker_t".

Get rid of extra unnecessary variables in "add_hostlist_table_data()".

Clean up indentation.

svn path=/trunk/; revision=11066