DCERPC problem reported by JBM and identified by Todd Sabin
Other protocol, not only SMB will populate pinfo->private_data
thus   checking for existence of non-NULL pinfo->private_data is not sufficient
to determine   we have SMB data and this is what it is.

Refactor the adding of salt/FID from lower layer protocols and generalize it.
Create a new dissector_handle specific for SMB so that we know that IFF we came in through that handle, then whatever pinfo->private_data is what we expect it to be.

svn path=/trunk/; revision=11129