Add "gtk/webbrowser.h" to declare functions from "gtk/webbrowser.c".
Rename "browser_open_program_file()" to "browser_open_data_file()", and
make it open files relative to the application's data directory, as
that's where data files such as HTMLized man pages would be put.  (That
happens to be the program directory on Windows, but it's a different
directory on UN*X - and you aren't guaranteed to be able to find the
program directory on UN*X by looking at argv[0] in any case.)  Move it
to "gtk/webbrowser.c".

Fix "filename2url()" to put "file://", not just "file:", in front of
pathnames on UN*X.

svn path=/trunk/; revision=11216