Add some comments, and expand on a comment (if "pcap_stats()" returns counts since the last call, not since the beginning of the capture, on Windows, that's a bug - and that bug *does* exist on some UN*Xes).
Clean up indentation.

Use the minimum snapshot length, not the maximum snapshot length, when
opening a network interface to keep track of how many packets are
arriving on it (that might cause packet data to be discarded rather than
buffering it, reducing the overhead of running all those captures from
which we don't actually grab any packets.

svn path=/trunk/; revision=11404