Add a color filter editor dialog, and hook it into the coloring rules dialog.
We still need a callback for the different display filter textboxes that
change the background color to red or green, depending on the validity
of the filter

We also need to implement the expression editor dialog.

Tweak the automatically generated hide_dialog routine to enable its owner.
This lets you disable the owner at init time, so the child remains in
focus no matter what.  Also have the hide_dialog routine activate the
owner window, which keeps Windows (XP, at least) from activating a random
window when the child goes away.

Other miscellaneous changes and fixes.

svn path=/branches/win32-native/; revision=11755