Add <toolbarseparator> elements.
Allow element to be hidden, and hook in the show/hide menu items under
the "View" menu.

Allow <toolbarbuttons> to have text labels.  (This isn't currently used.)

Hook in the "Time Display Format" menu items.

Move the code to reload files from gtk/file_dlg.c to file.c.  Use it to
hook in "View/Reload" and the reload toolbar button.

Hook in "View/Auto Scroll in Live Capture".  Add a note that this needs
to be synchronized with the capture and preference dialogs, both in the
native Win32 and GTK+ code.


svn path=/branches/win32-native/; revision=11822