Make <menulist> elements editable, and get rid of the <ethereal:combobox> element, since it's no longer needed.
Merge in the recent get_if_name() move from the trunk.

Add the full interface name, IP address, and link type selector to the
capture dialog.  Hook in the save file button.

Add OFN_NOCHANGEDIR to the list of flags that we pass to each instance of
GetOpenFileName() and GetSaveFileName().  This keeps them from changing the
current directory.  Other code (child capture process forking in particular)
depend on a stable working directory.


svn path=/branches/win32-native/; revision=11923