Add capture and filter selection dialogs.
Add syntax highlighting for display filters.  Getting this to work for
the main filter entry was a bit of a hack, and it still doesn't work
correctly.  Syntax checking hasn't been added to the file open dialog yet.

Allow listbox elements to handle double-clicks.

Add a routine to find an element in a window given its identifier.

Miscelleaneous fixes and cleanups.

svn path=/branches/win32-native/; revision=12031