Distinguish between the "data offset" field (offset of this burst within the file) and "burst offset" field (offset of this packet within the burst).
The burst header is not present if the SYS flag is set in the packet or
if the data offset field is non-zero.

Compute the offset, within a burst packet, of the data, as we advance
through the burst header - and don't advance through the burst header if
it's not present.

Properly display the fields in the "missing fragment list".

svn path=/trunk/; revision=12041