From Brian Caswell:
The ACL parser will attempt to decode as many ACE structures as are
specified in the ACL structure.  If the number of ACE structures is
sufficiently large with one of the ACE structures specifying a size of
0, then the ACL parser will parse that ACE structure repeatedly,
eventually causing a denial of service to Ethereal.

I've attached a diff against HEAD that corrects the problem.  The diff
also corrects a few decoding errors in the NT ACL & ACE structures.  A
pcap is attached that reproduces the problem.

svn path=/trunk/; revision=12706