add detection to TCP Analysis to detect and flag segments that will completely fill the window advertized from the other side.
I.e. when a segment is seen that would (as far as ethereal can tell from the ACKs it has seen in the other direction) fill the window completely.

It is similar to but not exactly the same as the XeroWindow detection since there are many instances where ZeroWindow detection would not work (i.e. an ACK where win==0    since many many situations occur where the window is full but no zerowindowack is ever generated)


Someone that has good english could, please, update the Wiki with this option.
It is very very useful to spot performance issues where the tcp window size is too small to accomodate the enmd-to-end latency.





svn path=/trunk/; revision=12774