Make the array for the "setup method" for RTP, RTCP, and T.38 conversations large enough to hold the maximum setup method size plus a trailing '\0'.  Make the maximum setup method size 7, so that when the trailing '\0' is included the total array length is a power of 2.  (The longest string currently used is "Skinny", which fits in 7 characters).
This fixes problems in the RTP and RTCP dissectors similar to the one
found in the T.38 dissector.

Undo the previous change to packet-t38.c, as it's now safe to store in
method[MAX_T38_SETUP_METHOD_SIZE], because the array now has
MAX_T38_SETUP_METHOD_SIZE+1 characters.

(Should we use "strlcpy()", and supply our own "strlcpy()" if the system
and/or C library doesn't supply it?  Its semantics are a bit cleaner
than those of the "strncpy()"/null-terminate idiom, perhaps making it
less likely that mistakes of this sort will be made.)

svn path=/trunk/; revision=12803