Set the eol-style and keywords properties.
Use "tvb_reported_length_remaining()", not "tvb_length_remaining()", in
the loop parsing the packet contents, so we throw an exception on a
short frame (to mark that it *is* a short frame).

Use "tvb_format_text()" for text strings, so we don't have a problem
with non-printable characters.

Use "ether_to_str()" to turn MAC addresses into strings.

Clean up indentation.

svn path=/trunk/; revision=13679