Handle the packet headeer for sFlow V5, put in a stub dissector for V5 samples, and give the URL of the V5 spec and some hints about how to handle the private enterprise code in those samples.
For subtrees that cover structures, make the tree cover the entire
structure, not just the first field.

Add the flow sample type and extended data type to the protocol tree.

Set the protocol column name to capitalize "sFlow" the conventional way,
and remove the redundant "sFlow" from the Info column.

svn path=/trunk/; revision=13744