Update the URL for draft-stewart-sctp-pktdrprep to the -02 draft.
Clean up indentation.

Have the main loop for dissecting chunks check that the chunk size is
large enough for the chunk header, and have it pass the chunk size,
minus the size of the chunk header, to dissectors for particular chunk
types.  Make those dissectors check that value to make sure it's large
enough for any fixed-length portion before subtracting the length of
that portion and using the result as a remaining data length.

svn path=/trunk/; revision=13924