Clean up white space.
Register "standard" and "lock_info_col" as obsolete preferences, so
they're silently ignored rather than triggering warnings.

Propagate changes to check for "dissect_ber_octet_string()" supplying a
null tvbuff pointer from the checkins to the generated dissector ("Do
not modify this file" means "Do not modify this file"!) to the tcap.cnf
file, so that they're not lost if the dissector is regenerated.  (Don't
propagate the checks for "tvb_new_subset()" returning null - it never
returns null.)

Pass -1 to "tvb_new_subset()" when appropriate.

When looping over packet data, check against the reported length, not
the captured length, so we indicate a short frame if we get stopped by
the snapshot length.

Get rid of unused variables.

svn path=/trunk/; revision=14658