Have separate fields for NFSv2, NFSv3, and NFSv4 return statuses, using the names in the v2, v3, and v4 specs for values, so that they display correctly and can be filtered on correctly.
Have "nfs.status" as a hidden field, with a name table giving just "OK"
and "ERR_" values, and having items for v2, v3, and v4 status values
(the only collision is that the v3 JUKEBOX error is the v4 DELAY error),
so that filters on "nfs.status" will still work, but will now be able to
filter symbolically on values from v2, v3, and v4.

Show 33 status values in the Info column and the summary entry using the
table of v3 status values and names, not the table of v2 status values
and names.

Fix some v3 dissectors to use the v3 return status field rather than the
v2 return status field.

svn path=/trunk/; revision=14831