From Ainsley Pereira:
If you have an RTP stream in which the first packet is duplicated, and has the marker bit set, then the RTP stream analysis will try to pad silence into a temporary file from 0 upto whatever the RTP timestamp is (which is enough to fill up my disk...). The  patch fixes this misbehaviour.

svn path=/trunk/; revision=14850