It works reasonably well with extensive testing so here goes :
tcp-graph has been modified to extract tcp data from a TAP instead of reading and parsing directly from the capture file.

This makes tcp graph work for any type of capture and for any transport.




In the future someone with access to captures with TCP over something over TCP
to add a dialog where the user can specify WHICH of the multiple TCP sessions to graph.



svn path=/trunk/; revision=14889