For packets flagged as Retransmission
calculate RTO as the delta between the retransmitted frame and the last previous frame seen for this session (in the same direction).

while this is technically not the RTO  this delta is in most cases more important/useful than the tru RTO anyway since this measure represents the amount of thiime that the link went idle  while waiting for an RTO.




It would be nice with a statistics tap for TCP where one couls see, seeion by session :
Length in time of the session.
Total bytes transferred
Number of retransmissions
Time spent waiting for an RTO
Time spent waiting for an RTO in % of the total time.




svn path=/trunk/; revision=14890