Fix the spelling of "transferred".
Fix the sizes of some items to match what's in the I-D and what's being
fetched from the packet.

Check for the length of items before doing *anything* with the item,
including fetching the data from the packet.  Just put a "bad length"
indication into the protocol tree; don't abort the dissection (we can
just move on to the next TLV).

svn path=/trunk/; revision=15254