We use heuristics to detect what MIGHT be onc-rpc packets.
as heuristics they are not 100% perfect, else they would not be heuristics.


IF the preference option "Dissect unknown RPC program numbers" are enabled,
add a sanity check to see that program_version is <=10
  (no one uses versions >10 ?)
to avoid trying to add 2 bilion different versions to the program->version linked list
in case we mistook the packet for ONC-RPC and the "version" field contains a huge random number.


This bug would only trigger a hang/crash IFF the option above is enabled.


svn path=/trunk/; revision=15290