Change a switch into an "if else if else if else" that fixes the "too large for switch" in HP-UX 10.20 .
The generated file has a lot of indentation changes due probably to a change in asn2eth.

The underlying reason for which I added this code (dissect_ber_integer() would not add as a "filterable" field if it is larger than 4 bytes, should be handled in dissect_ber_integer() ) remains there.

Should we change dissect_ber_integer() itself to make sure that a 5 byte unsigned integer which fits in 4 bytes gets added to the tree as "filterable"?


svn path=/trunk/; revision=15294