Get rid of unused variables.
Use "tvb_format_text()" and friends to format strings for display.

Don't throw an exception on errors if we can keep dissecting.

Use "PROTO_ITEM_SET_GENERATED()" to mark items as generated.

If an item covers real data, it's not generated.

Dissect WRITE FILEMARKS (6) (we already had a dissector, but it wasn't
being used).

svn path=/trunk/; revision=15341