Use "g_strdup_printf()" rather than formatting into a fixed-length buffer and "g_strdup()"ing it.
Use the nstime.c routines to compute time deltas and to add up times.

Don't add rows to the CList until a procedure gets a non-zero call
count, so only the ones with non-zero call counts are displayed (and
especially so that procedure numbers not corresponding to real
procedures aren't displayed!).

Ethereal requires 64-bit integer support, so don't bother checking
whether we have it.

Use the right format for 64-bit integers.

svn path=/trunk/; revision=15506