Add a "time_stat_init()" routine to initialize the fields of a "timestat_t".
Move "nstime_to_msec()" to "epan/nstime.c", as it has nothing to do with
a "timestat_t".

Use structure assignment when possible.

Fix the "addtime()" macro and use it in "time_stat_update()".

Use "timestat_t"s, and the routines to manipulate them, in the service
response time table code.

svn path=/trunk/; revision=15509