minor workflow enhancement, generate the HTML files before the pdf's and alike.
while working with the docbook sources, the html versions are a lot easier to look at than the pdf's. At least the Adobe Reader will lock the pdf file, so one has to close it to generate a next version, it's ugly to work with.

the html files can be viewed with e.g. Firefox while a new version is already generated, a simple reload will do the update then ...

svn path=/trunk/; revision=15625