The previous checkin for packet-isup.c was accidentally checked in along with some unrelated changes; it got rid of some calls to get IPv4/IPv6 addresses and add them to the protocol tree, replacing them with proto_tree_add_item() calls, and fixed up the length of one protocol tree item.
Note that we have (at least) three count 'em three different routines
for dissecting/displaying NSAPs (there might be more - I might have
missed some), and suggest that we might want to reduce that to one.

Update the URL for ICP values for IPv{4,6} addresses inside NSAPs.

Fix the offset in the call to add the IPv6 address part of an NSAP
containing such an address, and the length in the call adding the DSP
for an NSAP containing an IPv4 address.

Fix up indentation a bit.

svn path=/trunk/; revision=15766