Don't fill in a data structure for the packet while dissecting; we use almost none of the data - fill in only variables for what we need, and use proto_tree_add_item() in most cases.
Move what's left of the packet-winsrepl.h header into packet-winsrepl.c,
and get rid of the header.

Dissect the name flags field in detail, as per the Samba code.

We don't do any checks for whether the packet is a valid WINS
replication packet, so don't make the dissector a new-style dissector.

svn path=/trunk/; revision=15936