Add #defines of NDS protocol types, and use them in the value_string and switch statements.
Add some sanity checking for ptvcursor_advance calls.

proto_tree_add_item() works fine for IPv4 addresses; use it for them. 
(tvb_get_letohl() doesn't work fine for them, even though it appears to
do so on little-endian machines.)

Only set the "visible" flag on the protocol tree when we're sure we have
one, so that it gets set even if we create a temporary tree.

Catch exceptions when dissecting requests, and compute and save the
results of conditional tests before rethrowing the exception, so the
results get saved even if the dissection gets an exception.

svn path=/trunk/; revision=16288