in order to handle dcerpc over smb2 later and to make the heuristics stronger for determining what might and what might not be a pipe
add   TID tracking.   for all TreeConnect requests/resposnes seen,   store the name->tid mapping and other metadata.


as a freebee    the disswection of the tid in the ehader is aware of this table so when a tid value is dissected in the header and we known the name for this tid   then put it in an expansion below the tid.




svn path=/trunk/; revision=16483