Make the message length in the sync pipe message header 3 bytes of binary data, so the maximum message length can be up to 2^24-1.
Add a #define for that message size, and have it be 4096, for now, as
that was the size of the buffer used to read sync pipe messages.

Clean up white space.

When displaying an error message from the capture child, don't use it as
a format string - it could conceivably contain "%"s.  Instead, format it
with "%s".

svn path=/trunk/; revision=16541