Fix RFC number.
Use "tvb_reported_length_remaining()" for all length checks and values,
so we get the *actual* packet length, not the *captured* packet length
(it's a *feature* if it reports a "packet cut short" error if the packet
really *was* cut short by slicing).

Set subtree lengths to -1, so that if an exception is throws while
dissecting them, the length includes all the data.

Get rid of unused variables, and mark unused parameters.

In heuristic checks, give up if we don't have enough data to check.

svn path=/trunk/; revision=16599