Fix indentation.
Gate X and Y Offset are 16-bit integral numbers of degrees, not 32-bit
IEEE floating-point numbers, and the numbers in collision detection
segment definitions and collision detection volume definitions are
16-bit fixed-point numbers, not 32-bit IEEE floating-point numbers;
handle them as such - this fixes bug 605.

When displaying the packet length in the Info column, display the *real*
length, not the sliced length.

svn path=/trunk/; revision=16714